Refactor: split `lib-dns-resolver` out of `bin-resolved`
I'll need this to implement a `dig`-like command.
