+++ bin-resolved/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod fs_util;
 pub mod metrics;
-pub mod net_util;
-pub mod resolver;
