+++ lib-dns-resolver/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "dns-resolver"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-recursion = "1"
+bytes = "1"
+dns-types = { path = "../lib-dns-types" }
+priority-queue = "1"
+rand = "0.8.5"
+tokio = { version = "1", features = ["io-util", "net", "time"] }
+
+[dev-dependencies]
+criterion = "0.3.5"
+dns-types = { path = "../lib-dns-types", features = ["test-util"] }
+
+[[bench]]
+name = "cache"
+harness = false
