+++ lib-dns-resolver/src/lib.rs
@@ -15,7 +15,7 @@ use self::forwarding::resolve_forwarding;
 use self::metrics::Metrics;
 use self::nonrecursive::resolve_nonrecursive;
 use self::recursive::resolve_recursive;
-use self::util::ResolvedRecord;
+use self::util::types::ResolvedRecord;
 
 /// Maximum recursion depth.  Recursion is used to resolve CNAMEs, so
 /// a chain of CNAMEs longer than this cannot be resolved.
